+++ CHANGES.rst
@@ -5,6 +5,15 @@
   for services like ESO's and MAST's TAP, which do not use canonical
   prefixes while astropy.utils.xml ignores namespaces. [#323]
 
+
+1.3.1 (unreleased)
+==================
+
+- pyvo.dal warns on results with overflow status. [#329]
+
+- Allow session to be passed through in SSA and DataLink. [#327]
+
+
 1.3 (2022-02-19)
 ==================